+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/combat/TickBase.kt
@@ -35,7 +35,7 @@ object TickBase : Module("TickBase", Category.COMBAT) {
 
     private val rangeToAttack by floatRange("RangeToAttack", 3f..5f, 0f..10f)
 
-    private val forceGround by boolean("ForceGround", false)
+    private val onlyGround by boolean("OnlyGround", false)
     private val pauseAfterTick by int("PauseAfterTick", 0, 0..100)
     private val pauseOnFlag by boolean("PauseOnFlag", true)
 
@@ -90,7 +90,7 @@ object TickBase : Module("TickBase", Category.COMBAT) {
                 val tickDistance = tick.position.distanceTo(nearbyEnemy.positionVector)
 
                 (index to tick).takeIf {
-                    tickDistance < currentDistance && tickDistance in rangeToAttack && !tick.isCollidedHorizontally && (!forceGround || tick.onGround)
+                    tickDistance < currentDistance && tickDistance in rangeToAttack && !tick.isCollidedHorizontally && (!onlyGround || tick.onGround)
                 }
             }